+++ Petite/bus_lcd.h
@@ -51,7 +51,7 @@ extern s32 bus_lcd_write_data(DevLcdBusNode *node, u8 *data, u32 len);
 extern s32 bus_lcd_flush_data(DevLcdBusNode *node, u8 *data, u32 len);
 extern s32 bus_lcd_flush_wait(DevLcdBusNode *node);
 extern s32 bus_lcd_read_data(DevLcdBusNode *node, u8 *data, u32 len);
-extern s32 bus_lcd_write_cmd(DevLcdBusNode *node, u8 cmd);
+extern s32 bus_lcd_write_cmd(DevLcdBusNode *node, u16 cmd);
 extern s32 dev_lcdbus_register(const DevLcdBus *dev);
